Cleaned up seek tests a bit
