+++ src/Player/models/AudioPlayer.getDuration.test.ts
@@ -62,4 +62,39 @@ describe("AudioPlayer getDurationAsync", () => {
     expect(audioPlayer.getDurationAsync()).resolves.toEqual(123);
     audioPlayer.selectedTrack!.howl._emit("load");
   });
+
+  test("seekToTimestamp should change the current playback position", () => {
+    vi.spyOn(Howl.prototype, "state").mockImplementationOnce(() => "loaded");
+    vi.spyOn(Howl.prototype, "seek").mockReturnValue(40);
+    const track = audioPlayer.createTrack(trackDetails);
+    audioPlayer.selectedTrack = track;
+
+    const seekPosition = 30; // Seek to 30 seconds
+    const mockSeek = vi.spyOn(Howl.prototype, "seek");
+
+    audioPlayer.seekToTimestamp(seekPosition);
+
+    expect(mockSeek).toHaveBeenCalledWith(seekPosition);
+    mockSeek.mockRestore();
+  });
+
+  test("seekToTimestamp should not change the playback position if no track is loaded", () => {
+    const mockSeek = vi.spyOn(Howl.prototype, "seek");
+
+    audioPlayer.seekToTimestamp(30);
+
+    expect(mockSeek).not.toHaveBeenCalled();
+    mockSeek.mockRestore();
+  });
+
+  test("seekToTimestamp should return null if seek doesn't return a number", () => {
+    vi.spyOn(Howl.prototype, "state").mockImplementationOnce(() => "loaded");
+    const mockSeek = vi
+      .spyOn(Howl.prototype, "seek")
+      .mockReturnValue({} as Howl);
+    audioPlayer.seekToTimestamp(30);
+
+    expect(mockSeek).not.toHaveBeenCalled();
+    mockSeek.mockRestore();
+  });
 });
